backend/apps/accounts/urls.py
"""
Authentication URLs
"""
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView
from .views import RegisterView, LoginView, LogoutView, CurrentUserView, UpdateProfileView

app_name = 'accounts'

urlpatterns = [
    # Authentication
    path('register/', RegisterView.as_view(), name='register'),
    path('login/', LoginView.as_view(), name='login'),
    path('logout/', LogoutView.as_view(), name='logout'),

    # Token refresh
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    path('profile/', UpdateProfileView.as_view(), name='update_profile'),

    # Current user
    path('me/', CurrentUserView.as_view(), name='current_user'),
]