backend/requirements.txt
# Existing packages
Django==5.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
qrcode==7.4.2
unidecode==1.3.8

# ========== AI UCHUN (YANGI!) ==========
google-generativeai==0.3.2

# open ai dan
openai>=1.0.0
anthropic>=0.20.0

celery
redis

django-cors-headers

drf-yasg

psycopg2-binary

